refactor - changed file structure
