+++ src/main/java/com/example/SchedulEx/models/User.java
@@ -1,5 +1,6 @@
 package com.example.SchedulEx.models;
 
+import com.example.SchedulEx.helpers.PasswordHelper;
 import jakarta.persistence.*;
